Added own glock 18 model,sprinting, idle and run animation!
I have modeled my own glock 18 model in blender, im probably not going to use it in the final game but it works fine for testing. I have added sprinting and animations for idle and running. I olso did some more sorting to make it easier for me in the future.